flash: nor: scqspi: implement flash bank command
This commit adds the implementation of the `flash bank` command for the
SCQSPI flash driver.

The command initializes the driver’s private data structure
(`scqspi_flash_bank`) and parses the `io_base` address from the command
line.

The SPI SS line is set to a default value of 0x01 (Config memory 0).

This is a necessary step toward enabling probing and interaction with
the QSPI controller.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>